fix(RequestDecoratorFactory.java) Transform targetMethod to uppercase.
fix(GetRequestDecorator.java) URI is encoded when getting it from Exchange, and it gets encoded again when passing this URL through the UriComponentsBuilder resulting in query params like '%2520'. Here '%20' is the empty space character encoded but when it gets encoded again we end up with the encoded character of the percentage symbol '%25'. This could be added to the other request decorators too.